add functionality to pass data from new wallet creation to new wallet of savings type creation
When creating a new wallet, if the user selects to create a new savings type, navigate to the new savings wallet creation page. The data that was already filled out on the wallet page needs to be passed to the new page as well.

Use navigate to pass data, and use location to retrieve data